delete test project again. add new test
