--- Src/MoneyFox.Droid.Tests/MainActivity.cs
@@ -1,23 +1,0 @@
-﻿using System.Reflection;
-using Android.App;
-using Android.OS;
-using Xamarin.Android.NUnitLite;
-
-namespace MoneyFox.Droid.Tests
-{
-    [Activity(Label = "MoneyFox.Droid.Tests", MainLauncher = true, Icon = "@drawable/icon")]
-    public class MainActivity : TestSuiteActivity
-    {
-        protected override void OnCreate(Bundle bundle)
-        {
-            // tests can be inside the main assembly
-            AddTest(Assembly.GetExecutingAssembly());
-            // or in any reference assemblies
-            // AddTest (typeof (Your.Library.TestClass).Assembly);
-
-            // Once you called base.OnCreate(), you cannot add more assemblies.
-            base.OnCreate(bundle);
-        }
-    }
-}
-
