--- Src/MoneyFox.Droid.Tests/Resources/values/Strings.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="Hello">Hello World, Click Me!</string>
-    <string name="ApplicationName">MoneyFox.Droid.Tests</string>
-</resources>
