--- Src/MoneyFox.Droid.Tests/TestSample.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using NUnit.Framework;
-
-
-namespace MoneyFox.Droid.Tests
-{
-    [TestFixture]
-    public class TestsSample
-    {
-
-        [SetUp]
-        public void Setup() { }
-
-
-        [TearDown]
-        public void Tear() { }
-
-        [Test]
-        public void Pass()
-        {
-            Console.WriteLine("test1");
-            Assert.True(true);
-        }
-
-        [Test]
-        public void Fail()
-        {
-            Assert.False(true);
-        }
-
-        [Test]
-        [Ignore("another time")]
-        public void Ignore()
-        {
-            Assert.True(false);
-        }
-
-        [Test]
-        public void Inconclusive()
-        {
-            Assert.Inconclusive("Inconclusive");
-        }
-    }
-}
