--- Src/MoneyFox.Droid.Tests/project.json
@@ -1,14 +1,0 @@
-﻿{
-  "dependencies": {
-    "Microsoft.EntityFrameworkCore": "1.1.1",
-    "MvvmCross.Binding": "4.4.0",
-    "xunit": "2.3.0-beta1-build3642",
-    "xunit.runner.devices": "2.2.0-rc1"
-  },
-  "frameworks": {
-    "MonoAndroid,Version=v7.1": {}
-  },
-  "runtimes": {
-    "win": {}
-  }
-}
