--- Src/MoneyFox.Service.IntegrationTests/xunit.runner.json
@@ -1,5 +1,0 @@
-﻿{
-    "parallelizeAssembly": false,
-    "parallelizeTestCollections": false,
-    "methodDisplay": "method"
-}
